+++ config.tf
@@ -1,9 +1,6 @@
-resource "turbot_resource" "test" {
-  parent = "tmod:@turbot/turbot#/"
-  type = "tmod:@turbot/turbot#/resource/types/folder"
-  payload = <<EOF
-  {
-  "title": "provider_test2",
-  "description": "test resource2"
-}EOF
+
+resource "turbot_policy_setting" "test_policy" {
+  resource = "tmod:@turbot/turbot#/"
+  policy_type = "tmod:@turbot/aws#/policy/types/accountStack"
+  value = "Skip"
 }