+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+
+
+## 1.0.0-beta.2 (September 18, 2019)
+
+
+FEATURES:
+* **New Data Source:** `turbot_policy_value` ([#4](https://github.com/turbotio/terraform-provider-turbot/issues/4))
+
+
+BUGFIXES:
+
+* When searching for existing policy setting before creation, ignore default setting. ([#9](https://github.com/turbotio/terraform-provider-turbot/issues/9))
+ENHANCEMENTS:
+
+* resource/turbot_folder resource: remove error when creating folder with existing name and parent. This is permitted. ([#12](https://github.com/turbotio/terraform-provider-turbot/issues/12))
+* resource/turbot_policy_setting: add default value of "required" for precedence property.  ([#13](https://github.com/turbotio/terraform-provider-turbot/issues/13))
 ## 1.0.0-beta.1 (September 13, 2019)